+++ internal/runner/metrics_test.go
@@ -1,16 +1,16 @@
 package runner
 
 import (
-	"testing"
-
 	"encoding/json"
+	"testing"
+	"time"
 )
 
 //this file contains tests for the go
 
 func TestOutputMem(t *testing.T) {
 
-	jbuf, err := outputMem()
+	jbuf, err := OutputMem()
 
 	if err != nil {
 		t.Error(err)
@@ -32,13 +32,16 @@ func TestOutputMem(t *testing.T) {
 }
 
 func TestOutputCPU(t *testing.T) {
-	jbuf, err := outputCPU()
+	select {
+	case <-time.After(time.Duration(time.Second * 10)):
+	}
+	jbuf, err := OutputCPU()
 
 	if err != nil {
 		t.Error(err)
 	}
 
-	var vCPU interface{}
+	vCPU := map[string]interface{}{}
 	jsonErr := json.Unmarshal(jbuf, &vCPU)
 
 	if jsonErr != nil {
@@ -47,6 +50,39 @@ func TestOutputCPU(t *testing.T) {
 	if vCPU == nil {
 		t.Error("Missing CPU Utilization")
 	} else {
-		t.Logf("success, CPU: %v", vCPU)
+		t.Logf("success, CPU: %v", vCPU["cpuUtilization"])
+	}
+}
+
+func TestWrapJSON(t *testing.T) {
+	jbufM, _ := OutputMem()
+	jbufC, _ := OutputCPU()
+
+	var vMem map[string]interface{}
+	var vCPU map[string]interface{}
+
+	json.Unmarshal(jbufM, &vMem)
+	json.Unmarshal(jbufC, &vCPU)
+
+	t.Logf("%v", vCPU)
+
+	vJoin := map[string]map[string]interface{}{}
+
+	vJoin["currentMemory"] = vMem
+	vJoin["currentCPU"] = vCPU
+
+	vMetrics := map[string]map[string]map[string]interface{}{}
+
+	vMetrics["_metrics"] = vJoin
+
+	jsonMetrics, _ := json.Marshal(vMetrics)
+	//	t.Logf("%v", jsonMetrics)
+
+	var vCheck map[string]interface{}
+
+	json.Unmarshal(jsonMetrics, &vCheck)
+
+	if _, ok := vCheck["_metrics"]; ok {
+		t.Logf("%v", vCheck["_metrics"])
 	}
 }
